feat: refactor navbar structure and styles for improved responsiveness and maintainability; update color variables for consistency
